+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Sourcery now will not create empty files and will remove existing generated files with empty content if CLI flag `prune` is set to `true` (`false` by default)
 - Sourcery now will remove inline annotation comments from generated code.
 - Added `rethrows` property to `Method`
+- Allow duplicated annotations to be agregated into array
 
 ### Fixed